refactor(broker): add partitioning package
Move topology package into partitioning package
