+++ broker/src/main/java/io/camunda/zeebe/broker/partitioning/topology/TopologyPartitionListener.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.broker.clustering.topology;
+package io.camunda.zeebe.broker.partitioning.topology;
 
 import io.camunda.zeebe.protocol.impl.encoding.BrokerInfo;
